// src/libs/sdk/sdk-impl/sdk-types.ts
export interface InputFile {
  path: string;
  extension: string;
  srcPath?: string;
  getContents: () => Promise<string> | string;
}

export interface OutputFile {
  /**
   * relative to distDir
   */
  path: string;
  srcPath?: string;
  extension?: string;
  contents?: string;
  declaration?: boolean;
  errors?: Error[];
  raw?: boolean;
  skip?: boolean;
}

export type LoaderResult = OutputFile[] | undefined;

export type LoadFile = (input: InputFile) => LoaderResult | Promise<LoaderResult>;

interface LoaderOptions {
  ext?: "js" | "mjs" | "cjs" | "ts" | "mts" | "cts";
  format?: "cjs" | "esm";
  declaration?: boolean;
  esbuild?: CommonOptions;
  postcss?:
    | false
    | {
        nested?: false | PostcssNestedOptions;
        autoprefixer?: false | AutoprefixerOptions;
        cssnano?: false | CssnanoOptions;
        plugins?: PostcssPlugin[];
        processOptions?: Omit<PostcssProcessOptions, "from">;
      };
}

export interface LoaderContext {
  loadFile: LoadFile;
  options: LoaderOptions;
}

export type Loader = (
  input: InputFile,
  context: LoaderContext,
) => LoaderResult | Promise<LoaderResult>;

// should be the same as the loaders in loaders-mod.ts
type LoaderName = "js" | "vue" | "sass" | "postcss";

export type CreateLoaderOptions = {
  loaders?: (Loader | LoaderName)[];
} & LoaderOptions;

export type MkdistOptions = {
  rootDir?: string;
  srcDir?: string;
  pattern?: string | string[];
  globOptions?: GlobOptions;
  distDir?: string;
  cleanDist?: boolean;
  loaders?: (LoaderName | Loader)[];
  addRelativeDeclarationExtensions?: boolean;
  typescript?: {
    compilerOptions?: TSConfig["compilerOptions"];
  };
} & LoaderOptions;